Update README.md
This GitHub project is a collection of country flags recreated entirely using pure CSS. Each flag has been meticulously designed and implemented to demonstrate the power and versatility of CSS in creating visually appealing and scalable flag representations.

Sorry for the limited selection of flags at the moment. As a beginner, I'm continuously learning and honing my skills, and I will strive to create more country flags in the future. The more I learn, the more flags I will be able to create.